add selection in payment page
