+++ src/components/inference/inference-model-list.module.css
@@ -1,0 +1,121 @@
+/* Models — compact rows, each its own Card */
+.modelList {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.modelRow {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  width: 100%;
+  padding: 12px;
+  background: none;
+  border: none;
+  cursor: pointer;
+  color: inherit;
+  text-align: left;
+
+  .modelAvatar {
+    flex-shrink: 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 32px;
+    height: 32px;
+    border-radius: 8px;
+    overflow: hidden;
+    background: var(--background-glass-secondary);
+    font-weight: 700;
+    font-size: 14px;
+
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+  }
+
+  .modelIdentity {
+    display: flex;
+    flex-direction: column;
+    min-width: 0;
+    flex: 1;
+
+    .modelName {
+      font-weight: 600;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+    }
+
+    .modelSub {
+      font-size: 12px;
+      color: var(--text-secondary);
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+    }
+  }
+
+  .modelSpecs {
+    display: none;
+    flex-shrink: 0;
+    gap: 6px;
+
+    @media (min-width: 640px) {
+      display: flex;
+    }
+
+    .specChip {
+      font-size: 12px;
+    }
+  }
+
+  .chevron {
+    transition: transform 0.2s ease;
+
+    &.chevronOpen {
+      transform: rotate(180deg);
+    }
+  }
+}
+
+.paramsGrid {
+  display: grid;
+  gap: 12px 32px;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  margin: 0;
+  padding: 12px 24px 24px;
+
+  .paramItem {
+    display: grid;
+    grid-template-columns: auto 1fr;
+    align-items: baseline;
+    gap: 2px 12px;
+    padding-bottom: 8px;
+    border-bottom: 1px solid var(--border);
+
+    .paramLabel {
+      margin: 0;
+      font-size: 14px;
+      color: var(--text-secondary);
+    }
+
+    .paramValue {
+      margin: 0;
+      justify-self: end;
+      font-weight: 600;
+      text-align: right;
+    }
+
+    .paramFlag {
+      grid-column: 1 / -1;
+      font-family: var(--font-mono, monospace);
+      font-size: 12px;
+      color: var(--text-secondary);
+      opacity: 0.7;
+    }
+  }
+}
