+++ src/components/inference/payment-page.module.css
@@ -1,0 +1,7 @@
+.sectionHead {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: baseline;
+  justify-content: space-between;
+  gap: 8px;
+}
