+++ src/components/inference/payment-page.tsx
@@ -1,5 +1,7 @@
 import Card from '@/components/card/card';
 import Container from '@/components/container/container';
+import InferenceEnvironmentCard from '@/components/inference/inference-environment-card';
+import InferenceModelList, { fallbackParams, ServiceModel } from '@/components/inference/inference-model-list';
 import InferenceNavigation from '@/components/inference/inference-navigation';
 import InferencePayment from '@/components/inference/inference-payment';
 import InferenceStepper from '@/components/inference/inference-stepper';
@@ -7,14 +9,34 @@ import SectionTitle from '@/components/section-title/section-title';
 import { useInferenceContext } from '@/context/inference-context';
 import { getModelShortName } from '@/services/huggingface-service';
 import { InferenceFlowType } from '@/types/inference';
+import { formatDuration } from '@/utils/formatters';
 import { useParams } from 'next/navigation';
 import { useRouter } from 'next/router';
+import { useMemo } from 'react';
+import styles from './payment-page.module.css';
 
 const PaymentPage: React.FC<{ flowType: InferenceFlowType }> = ({ flowType }) => {
   const params = useParams<{ modelId?: string; templateId?: string }>();
   const router = useRouter();
-  const { selectedEnv, selectedToken, jobDurationSeconds, selectedModels, hydrateFromUrlFinished, buildSelectionQuery } =
-    useInferenceContext();
+  const {
+    selectedEnv,
+    selectedToken,
+    jobDurationSeconds,
+    selectedModels,
+    modelParamsByModel,
+    hydrateFromUrlFinished,
+    buildSelectionQuery,
+  } = useInferenceContext();
+
+  // Selected models paired with their committed launch params (falling back where none are set).
+  const models: ServiceModel[] = useMemo(
+    () =>
+      selectedModels.map((model) => ({
+        model,
+        params: modelParamsByModel[model.id] ?? fallbackParams(model.id),
+      })),
+    [selectedModels, modelParamsByModel]
+  );
 
   const goToPrevStep = () => {
     switch (flowType) {
@@ -53,22 +75,53 @@ const PaymentPage: React.FC<{ flowType: InferenceFlowType }> = ({ flowType }) =>
         contentBetween={<InferenceStepper currentStep="payment" flowType={flowType} />}
       />
       <div className="pageContentWrapper">
-        <Card direction="column" padding="md" radius="lg" shadow="black" spacing="md" variant="glass-shaded">
-          <h3>Payment</h3>
-          {!hydrateFromUrlFinished ? (
+        {!hydrateFromUrlFinished ? (
+          <Card direction="column" padding="md" radius="lg" shadow="black" spacing="md" variant="glass-shaded">
             <div className="textSecondary">Loading selection…</div>
-          ) : selectedEnv && selectedToken ? (
-            <InferencePayment
-              durationSeconds={jobDurationSeconds}
-              selectedEnv={selectedEnv}
-              selectedToken={selectedToken}
-            />
-          ) : (
-            <div className="textSecondary">Select an environment first.</div>
-          )}
-        </Card>
+          </Card>
+        ) : (
+          <Card direction="column" padding="md" radius="lg" shadow="black" spacing="md" variant="glass-shaded">
+            {/* Payment summary */}
+            {selectedEnv && selectedToken ? (
+              <InferencePayment
+                durationSeconds={jobDurationSeconds}
+                selectedEnv={selectedEnv}
+                selectedToken={selectedToken}
+              />
+            ) : (
+              <div className="textSecondary">Select an environment first.</div>
+            )}
 
-        <InferenceNavigation nextLabel="Pay & launch" onNext={goToNextStep} onPrev={goToPrevStep} />
+            {/* Models */}
+            {models.length > 0 && (
+              <>
+                <div className={styles.sectionHead}>
+                  <h3>Models</h3>
+                  <span className="textSecondary">{models.length} selected · expand for launch parameters</span>
+                </div>
+                <InferenceModelList models={models} />
+              </>
+            )}
+            {/* Environment */}
+            {selectedEnv && (
+              <>
+                <div className={styles.sectionHead}>
+                  <h3>Environment</h3>
+                  <span className="textSecondary">Running for {formatDuration(jobDurationSeconds)}</span>
+                </div>
+                <InferenceEnvironmentCard
+                  defaultToken={selectedToken?.address}
+                  durationSeconds={jobDurationSeconds}
+                  environment={selectedEnv.environment}
+                  gpuSelection={selectedEnv.gpuSelection}
+                  nodeInfo={selectedEnv.nodeInfo}
+                />
+              </>
+            )}
+          </Card>
+        )}
+
+        <InferenceNavigation hideSelection nextLabel="Pay & launch" onNext={goToNextStep} onPrev={goToPrevStep} />
       </div>
     </Container>
   );
